Add bulk_upsert functionality
A bulk upsert is essentially updating and inserting (in bulk) objects to
make a queryset contain the latest updates and new objects. This is more
efficient than `update_or_create` because it requires a single query to
check the existence of objects and requires a single query for the creation
of non-existing objects. Updates are performed individually.